+++ client/src/App.js
@@ -138,19 +138,19 @@ function App() {
             <Route path="/stats" element={<Stats />} />
             <Route path="/alerts" element={<Alerts />} />
 
-            <Route path="/todo" element={<Todos />} />
-            <Route path="/todo/:id" element={<TodoUpdate />} />
-            <Route path="/todo/create" element={<TodoCreate /> } />
+            <Route path="/todos" element={<Todos />} />
+            <Route path="/todos/:id" element={<TodoUpdate />} />
+            <Route path="/todos/create" element={<TodoCreate /> } />
 
-            <Route path="/system" element={<Systems />} />
-            <Route path="/system/:id" element={<SystemUpdate />} />
-            <Route path="/system/create" element={<SystemCreate />} />
+            <Route path="/systems" element={<Systems />} />
+            <Route path="/systems/:id" element={<SystemUpdate />} />
+            <Route path="/systems/create" element={<SystemCreate />} />
 
             <Route path="/light/create" element={<LightCreate /> } />
 
-            <Route path="/plant" element={<Plants />} />
+            <Route path="/plants" element={<Plants />} />
             <Route path="/plants/:id" element={<PlantUpdate />} />
-            <Route path="/plant/create" element={<PlantCreate />} />
+            <Route path="/plants/create" element={<PlantCreate />} />
 
             <Route path="/type/create" element={<TypeCreate /> } />
