+++ server/routes/__init__.py
@@ -1,0 +1,1 @@
+from .system_routes import bp as system_bp
